Divided `market.py` into two separate files. - Created the `PriceOverview` class, which will encapsulate all actions related to the `/priceoverview/` endpoint. - Adapted the `Market` class to support multiple-inheritance, in order to add other endpoints in the same manner. - Adapted `__init__.py` for the newly implemented changes.
